ml-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pfinance-ml-service"
version = "0.1.0"
description = "Self-hosted ML pipeline for document extraction"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "paddleocr>=2.9.0",
    "paddlepaddle>=2.6.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
modal = [
    "modal>=0.57.0",
]
gpu = [
    "paddlepaddle-gpu>=2.6.0",
]

[project.scripts]
ml-service = "src.main:run"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
